fix reactivity errors by creating a separate mutable object for edit form
